# Improve implicits resolution performance