Recently I better understood implicit prioritization rules and how to improve compile-time performance by just stacking them differently.

This attempt shows compilation time improvement up to 30%, measured on our own test suite.

The table below shows results of `chimneyJVM/test` prepended by `chimneyJVM/clean` every time on my computer.

|             | avg on unwarmed compiler  | avg on warmed compiler |
| ------ | ------------- | ------------- |
| before |  ~44s            | ~25s                 |
| after    |  ~34s            | ~17s                  |